refactor: complete blog optimization plan
- Item 6: Reduced !important from ~50 to 11 in globals.css (-78%)
- Item 10: Reorganized components dir (30 to 2 top-level files)
- Item 7: Evaluated lib modules (no change needed)
- Item 12: Verified N/A (recharts only in admin)

All 12 optimization items now completed/evaluated.
