+++ src/app/[locale]/gallery/map/page.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import { listGalleryImagesWithLocation } from "@/lib/gallery";
-import { GalleryMapWrapper } from "@/components/gallery-map-wrapper";
+import { GalleryMapWrapper } from "@/components/gallery/gallery-map-wrapper";
 
 // ISR: Revalidate every 5 minutes for gallery map data
 export const revalidate = 300;
