+++ src/app/[locale]/posts/[slug]/page.tsx
@@ -7,8 +7,8 @@ import remarkGfm from "remark-gfm";
 import { PostLocale, PostStatus } from "@prisma/client";
 import prisma from "@/lib/prisma";
 import { generateBlogPostingSchema, generateAlternateLinks } from "@/lib/seo";
-import { LikeButton } from "@/components/like-button";
-import { LanguageSwitcher } from "@/components/language-switcher";
+import { LikeButton } from "@/components/shared/like-button";
+import { LanguageSwitcher } from "@/components/layout/language-switcher";
 import { safeJsonLd } from "@/lib/safe-json-ld";
 import { cache } from "react";
 import { Container } from "@/components/ui/container";
