+++ src/components/auth/index.ts
@@ -1,0 +1,4 @@
+// Auth components barrel export
+export { AuthHeader } from "./auth-header";
+export { EmailLoginForm } from "./email-login-form";
+export { LoginButtons } from "./login-buttons";
