+++ src/components/gallery/gallery-map-wrapper.tsx
@@ -3,7 +3,7 @@
 import dynamic from "next/dynamic";
 import type { GalleryLocationImage } from "@/lib/gallery";
 
-const GalleryMap = dynamic(() => import("@/components/gallery-map").then((mod) => mod.GalleryMap), {
+const GalleryMap = dynamic(() => import("@/components/gallery/gallery-map").then((mod) => mod.GalleryMap), {
   ssr: false,
   loading: () => (
     <div className="flex h-[600px] items-center justify-center rounded-lg border border-stone-200 bg-stone-50 dark:border-stone-800 dark:bg-stone-900">
