+++ src/components/gallery/gallery-map.tsx
