+++ src/components/gallery/index.ts
@@ -1,0 +1,7 @@
+// Gallery components barrel export
+export { GalleryCard } from "./gallery-card";
+export { GalleryCategoryTabs } from "./gallery-category-tabs";
+export { GalleryGrid } from "./gallery-grid";
+export { GalleryMap } from "./gallery-map";
+export { GalleryMapWrapper } from "./gallery-map-wrapper";
+export { GalleryMasonry } from "./gallery-masonry";
