+++ src/components/layout/global-language-switcher.tsx
