+++ src/components/layout/index.ts
@@ -1,0 +1,10 @@
+// Layout components barrel export
+export { MainNav } from "./main-nav";
+export { Footer } from "./footer";
+export { ThemeProvider } from "./theme-provider";
+export { ThemeToggle } from "./theme-toggle";
+export { LanguageSwitcher } from "./language-switcher";
+export { GlobalLanguageSwitcher } from "./global-language-switcher";
+export { GlobalSearchProvider } from "./global-search-provider";
+export { HtmlLangSync } from "./html-lang-sync";
+export { SessionProvider } from "./session-provider";
