+++ src/components/photo-viewer/components/photo-viewer-toolbar.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { localePath } from "@/lib/locale-path";
-import { ThemeToggle } from "@/components/theme-toggle";
+import { ThemeToggle } from "@/components/layout/theme-toggle";
 import type { PhotoViewerToolbarProps } from "../types";
 
 export function PhotoViewerToolbar({
