+++ src/components/photo-viewer/photo-metadata-panel.tsx
