+++ src/components/shared/live-photo-player.tsx
