+++ src/components/shared/shuffle-grid-stats.tsx
