+++ src/components/zhi/post-card.tsx
@@ -1,99 +1,71 @@
-"use client";
-
-import React from "react";
-import { useRouter } from "next/navigation";
-import { useSession } from "next-auth/react";
-import { Heart } from "lucide-react";
 import Image from "next/image";
-
-interface BlogPost {
-  id: string;
-  title: string;
-  excerpt: string;
-  category: string;
-  date: string;
-  readTime: string;
-  imageUrl?: string;
-  tags: string[];
-  likes: number;
-}
+import Link from "next/link";
+import type { PublicPost } from "@/lib/posts";
+import { toOptimizedImageUrl } from "@/lib/image-proxy";
+import { localePath } from "@/lib/locale-path";
 
 interface PostCardProps {
-  post: BlogPost;
-  onClick?: () => void;
-  onLike?: (id: string) => void;
+  post: PublicPost;
+  locale?: "zh" | "en";
 }
 
-export function ZhiPostCard({ post, onClick, onLike }: PostCardProps) {
-  const { data: session } = useSession();
-  const router = useRouter();
-
-  const handleLike = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    if (!session) {
-      router.push("/login");
-      return;
-    }
-    onLike?.(post.id);
-  };
+export function PostCard({ post, locale = "zh" }: PostCardProps) {
+  const cover = toOptimizedImageUrl(post.coverImagePath) ?? "/images/placeholder-cover.svg";
+  const formatted = post.publishedAt
+    ? new Intl.DateTimeFormat(locale === "zh" ? "zh-CN" : "en-US", {
+      year: "numeric",
+      month: "long",
+      day: "numeric",
+    }).format(new Date(post.publishedAt))
+    : locale === "zh"
+      ? "草稿"
+      : "Draft";
 
   return (
-    <div className="group mb-8 cursor-pointer break-inside-avoid" onClick={onClick}>
-      <article className="border-b-2 border-stone-100 bg-white pb-6 transition-colors duration-300 hover:border-stone-300 dark:border-[#27272a] dark:bg-transparent dark:hover:border-stone-600">
-        {/* Image - Plain and sharp, full color */}
-        {post.imageUrl && (
-          <div className="relative mb-5 aspect-video w-full overflow-hidden rounded-sm">
-            <Image
-              src={post.imageUrl}
-              alt={post.title}
-              fill
-              sizes="(max-width: 768px) 100vw, (max-width: 1024px) 50vw, 384px"
-              className="object-cover transition-transform duration-700 group-hover:scale-105"
-              quality={80}
-            />
-            <div className="absolute top-3 left-3 z-10 border border-stone-200 bg-white px-3 py-1 text-[10px] font-bold uppercase tracking-widest dark:border-[#2a2a2e] dark:bg-[#141416]">
-              {post.category}
-            </div>
-          </div>
-        )}
-
-        <div className="px-3">
-          {!post.imageUrl && (
-            <div className="mb-3 text-[10px] font-bold uppercase tracking-widest text-stone-400">
-              {post.category}
+    <article className="group overflow-hidden border border-stone-200 bg-white transition dark:border-stone-800 dark:bg-stone-900">
+      <div className="relative aspect-[16/9] overflow-hidden bg-stone-100 dark:bg-stone-800">
+        <Image
+          src={cover}
+          alt={post.title}
+          fill
+          sizes="(max-width: 768px) 100vw, 50vw"
+          className="object-cover transition duration-500 group-hover:scale-105"
+        />
+      </div>
+      <div className="flex flex-col gap-6 p-8">
+        <div className="flex flex-wrap items-center gap-3 text-xs text-stone-500 dark:text-stone-400">
+          <time dateTime={post.publishedAt ?? ""}>{formatted}</time>
+          {post.tags.length > 0 && (
+            <div className="flex flex-wrap gap-2">
+              {post.tags.map((tag) => (
+                <span
+                  key={`${post.id}-${tag}`}
+                  className="font-medium text-stone-700 dark:text-stone-300"
+                >
+                  #{tag}
+                </span>
+              ))}
             </div>
           )}
-
-          <h3 className="mb-3 font-serif text-xl font-bold leading-tight text-stone-900 transition-colors group-hover:text-stone-600 md:text-2xl dark:text-stone-100 dark:group-hover:text-stone-300">
-            {post.title}
-          </h3>
-
-          <p className="mb-4 line-clamp-3 font-serif text-sm leading-relaxed text-stone-600 dark:text-stone-400">
-            {post.excerpt}
-          </p>
-
-          <div className="flex items-center justify-between font-sans text-xs text-stone-400">
-            <div className="flex items-center gap-3">
-              <span>{post.date}</span>
-              <span className="h-1 w-1 rounded-full bg-stone-300"></span>
-              <span>{post.readTime}</span>
-            </div>
-
-            <div className="flex items-center gap-4">
-              <button
-                onClick={handleLike}
-                className={`flex items-center gap-1 transition-colors hover:text-rose-500 ${post.likes > 0 ? "text-rose-500" : ""
-                  }`}
-              >
-                <Heart size={14} className={post.likes > 0 ? "fill-current" : ""} />
-                <span>{post.likes}</span>
-              </button>
-            </div>
-          </div>
         </div>
-      </article>
-    </div>
+        <h3 className="text-xl leading-snug font-semibold text-stone-900 dark:text-stone-50">
+          <Link href={localePath(locale, `/posts/${encodeURIComponent(post.slug)}`)}>
+            {post.title}
+          </Link>
+        </h3>
+        <p className="line-clamp-3 text-base leading-loose text-stone-600 dark:text-stone-400">
+          {post.excerpt}
+        </p>
+        <Link
+          href={localePath(locale, `/posts/${encodeURIComponent(post.slug)}`)}
+          className="text-sm font-medium text-stone-900 underline underline-offset-4 hover:text-stone-600 dark:text-stone-100"
+        >
+          {locale === "zh" ? "阅读全文" : "Read more"}
+        </Link>
+      </div>
+    </article>
   );
 }
 
-export default ZhiPostCard;
+// Alias for backwards compatibility
+export { PostCard as ZhiPostCard };
